third_party/blink/web_tests/external/wpt/html/cross-origin-opener-policy/reporting/tentative/access-to-noopener-page-from-no-coop-ro.https-expected.txt

This is a testharness.js-based test.
[FAIL] access-to-coop-page-from-opener, noopener-allow-popups
  assert_equals: expected (string) "coop" but got (undefined) undefined
[FAIL] access-to-coop-page-from-opener, noopener-allow-popups + redirect
  assert_equals: expected (string) "coop" but got (undefined) undefined
Harness: the test ran to completion.
